% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Recorder.R
\name{recorder}
\alias{recorder}
\title{Recorder Function}
\usage{
recorder(tutorial_id, tutorial_version, user_id, event, data)
}
\arguments{
\item{tutorial_id}{}

\item{tutorial_version}{}

\item{user_id}{}

\item{event}{}

\item{data}{}
}
\description{
Recorder Function
}
