% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{update_ENSC311}
\alias{update_ENSC311}
\title{Update the RBasics package}
\usage{
update_ENSC311()
}
\description{
Automatically installs the latest version of RBasics from github.
}
\details{
Simply run "update_ENSC311()" to get the latest version of the package.
}
